Combine login, register, and logout API endpoints
