+++ pages/login.js
@@ -48,7 +48,7 @@ export default function Login() {
 
             try {
               mutateUser(
-                await fetchJson("/api/login", {
+                await fetchJson("/api/auth", {
                   method: "POST",
                   headers: { "Content-Type": "application/json" },
                   body: JSON.stringify(body),
